refactor: console.log 삭제
